+++ package.json
@@ -14,6 +14,7 @@
   },
   "scripts": {
     "test": "jest",
+    "tdd": "npm run test -- --watch",
     "test-ci": "jest --coverage --ci && codecov",
     "dev": "start-storybook -p 6006 -s ./static",
     "lint:js": "eslint --ext=.tsx --ext=.ts src",
