+++ src/Molecules/Button/Button.Default.stories.tsx
@@ -1,6 +1,5 @@
 import React, { useState, useRef } from 'react';
 import { action } from '@storybook/addon-actions';
-import { HashRouter } from 'react-router-dom';
 import MD from 'react-markdown';
 import docs from './Button.md';
 
@@ -407,172 +406,4 @@ export const buttonsComposedInAGroup = () => (
 
 buttonsComposedInAGroup.story = {
   name: 'Buttons composed in a group',
-};
-
-export const linkLookingLikeAButton = () => (
-  <Display
-    items={[
-      {
-        title: 'Reset',
-        component: (
-          <HashRouter>
-            <Button to="route1">To Route1</Button>
-          </HashRouter>
-        ),
-      },
-      {
-        title: 'Submit',
-        component: (
-          <HashRouter>
-            <Button to="route2" variant="secondary">
-              To Route2
-            </Button>
-          </HashRouter>
-        ),
-      },
-    ]}
-  />
-);
-
-linkLookingLikeAButton.story = {
-  name: 'Link looking like a button',
-};
-
-export const linkLookingLikeAButtonThatIsFullWidth = () => (
-  <HashRouter>
-    <Button to="/overview" onClick={action('clicked')} fullWidth>
-      Button
-    </Button>
-  </HashRouter>
-);
-
-linkLookingLikeAButtonThatIsFullWidth.story = {
-  name: 'Link looking like a button that is full width',
-};
-
-export const linksLookingLikeButtonsComposedInAGroup = () => (
-  <HashRouter>
-    <Flexbox container gutter={2}>
-      <Flexbox item flex="1">
-        <Button to="route1" onClick={action('submit')} fullWidth>
-          Submit
-        </Button>
-      </Flexbox>
-      <Flexbox item flex="1">
-        <Button to="route2" onClick={action('reset')} variant="secondary" fullWidth>
-          Reset
-        </Button>
-      </Flexbox>
-    </Flexbox>
-  </HashRouter>
-);
-
-linksLookingLikeButtonsComposedInAGroup.story = {
-  name: 'Links looking like buttons composed in a group',
-};
-
-export const linkLookingLikeAButtonWithRelAsNofollow = () => (
-  <HashRouter>
-    <Button to="/overview" rel="nofollow" onClick={action('clicked')} fullWidth>
-      Button
-    </Button>
-  </HashRouter>
-);
-
-linkLookingLikeAButtonWithRelAsNofollow.story = {
-  name: 'Link looking like a button with rel as nofollow',
-};
-
-export const externalLinkLookingLikeAButton = () => (
-  <HashRouter>
-    <Button to="//www.google.com" external rel="nofollow" onClick={action('clicked')} fullWidth>
-      Button
-    </Button>
-  </HashRouter>
-);
-
-externalLinkLookingLikeAButton.story = {
-  name: 'External link looking like a button',
-};
-
-export const disabledLinkLookingLikeAButton = () => (
-  <HashRouter>
-    <Button
-      fullWidth
-      onClick={action('clicked')}
-      size="m"
-      to="/overview"
-      variant="secondary"
-      disabled
-    >
-      Button
-    </Button>
-  </HashRouter>
-);
-
-disabledLinkLookingLikeAButton.story = {
-  name: 'Disabled link looking like a button',
-};
-
-export const linkLookingLikeAButtonWithWithColors = () => (
-  <HashRouter>
-    <>
-      <Display
-        title="Primary"
-        items={[
-          {
-            title: 't => t.color.cta',
-            component: (
-              <Button to="/overview" color={t => t.color.cta} onClick={action('clicked')}>
-                Button
-              </Button>
-            ),
-          },
-          {
-            title: 't => t.color.negative',
-            component: (
-              <Button to="/overview" color={t => t.color.negative} onClick={action('clicked')}>
-                Button
-              </Button>
-            ),
-          },
-        ]}
-      />
-      <Display
-        title="Secondary"
-        items={[
-          {
-            title: 't => t.color.cta',
-            component: (
-              <Button
-                variant="secondary"
-                to="/overview"
-                color={t => t.color.cta}
-                onClick={action('clicked')}
-              >
-                Button
-              </Button>
-            ),
-          },
-          {
-            title: 't => t.color.negative',
-            component: (
-              <Button
-                variant="secondary"
-                to="/overview"
-                color={t => t.color.negative}
-                onClick={action('clicked')}
-              >
-                Button
-              </Button>
-            ),
-          },
-        ]}
-      />
-    </>
-  </HashRouter>
-);
-
-linkLookingLikeAButtonWithWithColors.story = {
-  name: 'Link looking like a button with with colors',
 };
