+++ src/Molecules/Button/Button.styles.tsx
@@ -34,21 +34,21 @@ const shared = css<InnerProps>`
   box-sizing: border-box;
   align-items: center;
   justify-content: center;
-  ${p => (p.fullWidth ? `display: flex; width: 100%;` : `display: inline-flex;`)}
+  ${p => (p.$fullWidth ? `display: flex; width: 100%;` : `display: inline-flex;`)}
 `;
 
 const minHeight = css<InnerProps>`
   ${p =>
-    p.size === 'm' || p.size === 'l'
-      ? `min-height: ${p.theme.spacing.unit(HEIGHT[p.size])}px;`
+    p.$size === 'm' || p.$size === 'l'
+      ? `min-height: ${p.theme.spacing.unit(HEIGHT[p.$size])}px;`
       : ''}
 `;
 
 const padding = css<InnerProps>`
   ${p => `
     padding:
-      ${p.theme.spacing.unit(PADDING_VERTICAL[p.size])}px
-      ${p.theme.spacing.unit(PADDING_HORIZONTAL[p.size])}px;
+      ${p.theme.spacing.unit(PADDING_VERTICAL[p.$size])}px
+      ${p.theme.spacing.unit(PADDING_HORIZONTAL[p.$size])}px;
   `}
 `;
 
@@ -57,7 +57,7 @@ export const primaryStyles = css<InnerProps>`
   ${padding}
   ${minHeight}
   background-color: ${p => {
-    const customColor = p.colorFn && p.colorFn(p.theme);
+    const customColor = p.$colorFn && p.$colorFn(p.theme);
     const background = customColor || p.theme.color.cta;
 
     return p.disabled ? p.theme.color.disabledBackground : background;
@@ -66,7 +66,7 @@ export const primaryStyles = css<InnerProps>`
   ${getBorder('transparent')}
 
   ${p => {
-    const customColor = p.colorFn && p.colorFn(p.theme);
+    const customColor = p.$colorFn && p.$colorFn(p.theme);
     const background = customColor || p.theme.color.cta;
 
     return p.disabled
@@ -90,7 +90,7 @@ export const secondaryStyles = css<InnerProps>`
     p.disabled ? p.theme.color.disabledBackground : p.theme.color.buttonSecondaryBackground};
 
   ${p => {
-    const customColor = p.colorFn && p.colorFn(p.theme);
+    const customColor = p.$colorFn && p.$colorFn(p.theme);
     const color = customColor || p.theme.color.cta;
 
     return `
@@ -128,7 +128,7 @@ export const neutralStyles = css<InnerProps>`
   background-color: transparent;
 
   ${p => {
-    const color = p.colorFn && p.colorFn(p.theme);
+    const color = p.$colorFn && p.$colorFn(p.theme);
 
     return `
       color: ${p.disabled ? p.theme.color.disabledText : color || p.theme.color.text}
